REFS: #243 and fixes #243(Changes in Myntra Project React)

+++ Modules/Project/Nihar Sodhaparmar/React/myntra/src/admin/pages/BrandsPage.js
@@ -91,20 +91,21 @@ export default function BrandsPage(props) {
   };
 
   const deleteBrand = async (id) => {
-    try {
-      setLoading(true);
-      await BrandService.deleteBrand(id, getToken());
-      getData();
-      setLoading(false);
-    } catch (error) {
-      console.error(error);
-      if (error.response.status === 401) {
-        removeUserSession();
-        props.history.push("/login");
-      } else if (error.response.status === 409) {
-        alert(error.response.data);
+    if (window.confirm("Are you sure, you want to delete?")) {
+      try {
+        setLoading(true);
+        await BrandService.deleteBrand(id, getToken());
+        getData();
+        setLoading(false);
+      } catch (error) {
+        if (error.response.status === 401) {
+          removeUserSession();
+          props.history.push("/login");
+        } else if (error.response.status === 409) {
+          alert(error.response.data);
+        }
+        setLoading(false);
       }
-      setLoading(false);
     }
   };
